feat: :construction: request manager and request type
