+++ .gitignore
@@ -1,1 +1,3 @@
 .venv
+.vscode
+__pycache__
