+++ LimAPI/Types.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class Request(BaseModel):
+    method: str
