+++ README.md
@@ -1,0 +1,6 @@
+# Легковесный веб-фреймворк для создания API
+
+## Основые особенности
+- Простота работы с инструментом
+- Настройка авторизации в две строчки
+- Автопагинация
